Refactor API and DB adapters, implement new API client structure, and update course domain serialization

+++ src/domain/course.rs
@@ -1,8 +1,8 @@
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Deserialize)]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct Course {
-    pub id: i32,
+    pub id: i64,
     pub fullname: String,
     // category: String,
     pub completed: Option<bool>,
